+++ lib/src/core/engine/analytics/capturer/view_story_event_capturer.dart
@@ -1,0 +1,10 @@
+import 'package:amity_sdk/src/core/engine/analytics/analytics_event_sync_priority.dart';
+import 'package:amity_sdk/src/core/engine/analytics/capturer/analytics_event_capturer.dart';
+
+class ViewStoryEventCapturer extends AnalyticsEventCapturer {
+  ViewStoryEventCapturer()
+      : super(
+            contentType: "story",
+            eventType: "view",
+            syncPriority: AnalyticsEventSyncPriority.HIGH);
+}