+++ lib/src/core/socket/event/mqtt_event_listeners.dart
@@ -1,8 +1,10 @@
 import 'package:amity_sdk/src/core/socket/event/event.dart';
 import 'package:amity_sdk/src/core/socket/event/post/post_approved_event_listener.dart';
 import 'package:amity_sdk/src/core/socket/event/post/post_declined_event_listener.dart';
+import 'package:amity_sdk/src/core/socket/event/story/story_reaction_added_event_listerner.dart';
 import 'package:amity_sdk/src/core/socket/event/story/story_create_event_listener.dart';
 import 'package:amity_sdk/src/core/socket/event/story/story_delete_event_listener.dart';
+import 'package:amity_sdk/src/core/socket/event/story/story_reaction_removed_event_listener.dart';
 import 'package:collection/collection.dart' show IterableExtension;
 
 class MqttEventListeners {
@@ -46,6 +48,9 @@ class MqttEventListeners {
     // Stories 
     StoryCreateEventListener(),
     StoryDeleteEventListener(),
+    StoryReactionAddedEventListener(),
+    StoryReactionRemovedEventListener(),
+    
 
   ]);